[BUGFIX] Image size calculation should be relative to PE base

+++ scanners/artefact_scanner.h
@@ -62,7 +62,7 @@ public:
 	size_t ntFileHdrsOffset; //offset from the regionStart
 	size_t secHdrsOffset; //offset from the regionStart
 	size_t secCount;
-	DWORD calculatedImgSize;
+	size_t calculatedImgSize;
 };
 
 class ArtefactScanReport : public MemPageScanReport
